% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_attendance.R
\name{check_attendance}
\alias{check_attendance}
\title{Check Attendance Records}
\usage{
check_attendance(check = "times")
}
\arguments{
\item{check}{One of "attendee", "date", "time_in", "time_out", "times", "location" Check looks for records with issues
in the corresponding column.  times checks that time_out occurs after time_in}
}
\value{
return tables of attendance records with issues
}
\description{
Check Attendance Records
}
\examples{
check_attendance()
}
