% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_contacts.R
\name{check_contacts}
\alias{check_contacts}
\title{Check Contacts}
\usage{
check_contacts(check = "state")
}
\arguments{
\item{check}{One of "phone", "email", "street", "state", nl", "full_name", "zip", or "solicit" check corresponding column.}
}
\value{
a flextable with the check of the requested column
}
\description{
Check Contacts
}
\examples{
check_contacts()
check_contacts("phone")
}
