% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicate_contacts.R
\name{check_duplicate_contacts}
\alias{check_duplicate_contacts}
\title{Check for Duplicate Contacts}
\usage{
check_duplicate_contacts(closeness = 3)
}
\arguments{
\item{closeness}{contacts with full names within Levenshtein closeness to include on report}
}
\value{
a flextable showing closeness of contact full names to other contacts full names
}
\description{
Contacts are checked for closeness measured by Levenshtein distance on full name. 0 indicates two full names are
identical,
1 indicates a single character edit (addition, deletion, or change) will make two full names identical.
}
\examples{
check_duplicate_contacts() 
check_duplicate_contacts(4) # relax the closeness for a longer list
}
