% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicate_members.R
\name{check_duplicate_members}
\alias{check_duplicate_members}
\title{Check for Duplicate Members}
\usage{
check_duplicate_members(closeness = 3)
}
\arguments{
\item{closeness}{members with names within Levenshtein closeness to include on report}
}
\value{
a flextable showing closeness of member last name, first name to other member's last name, first name
}
\description{
Members are checked for closeness measured by Levenshtein distance on last name, first name. 0 indicates two names are
identical,
1 indicates a single character edit (addition, deletion, or change) will make two names identical.
}
\examples{
check_duplicate_members()
check_duplicate_members(4) # relax the closeness for a longer list
}
