% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_employment.R
\name{check_employment}
\alias{check_employment}
\title{Check Employment}
\usage{
check_employment(check = "member")
}
\arguments{
\item{check}{The employment check to be run, one of member, employer, start, support, pay, or hours}
}
\value{
a flextable showing checks for the value specified
}
\description{
Check Employment
}
\examples{
check_employment()
check_employment("hours")
}
