% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gifts.R
\name{check_gifts}
\alias{check_gifts}
\title{Check Gifts}
\usage{
check_gifts(check = "donor")
}
\arguments{
\item{check}{The check to be performed, one of: donor, restriction, anonymous, amount, campaign, date, method, restricted, or thank_you}
}
\value{
a flextable with the requested gift checks
}
\description{
Check Gifts
}
\examples{
check_gifts()
check_gifts("restriction")
}
