% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_goals.R
\name{check_goals}
\alias{check_goals}
\title{Check Goals}
\usage{
check_goals(check = "lapsed_list")
}
\arguments{
\item{check}{Item to check.  Choices are: staff, member, text, lapsed_table, and lapsed_list}
}
\value{
a flextable with the checks
}
\description{
Check Goals
}
\examples{
check_goals()
check_goals("lapsed_list") # a list of the lapsed goals
check_goals("lapsed_table") # a table counting lapsed goals by staff member
}
