% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_members.R
\name{check_members}
\alias{check_members}
\title{Check Member Info}
\usage{
check_members(check = "contacts", active.only = T)
}
\arguments{
\item{check}{What to check.  Choices are "contacts", "emergency", "benefits", "demographics",
"housing", "education", "employment", medical", "referral", "admin", and "attendance"}

\item{active.only}{If TRUE, only active members are included}
}
\value{
a flextable of member information
}
\description{
Produce a table a member information
}
\examples{
check_members()
check_members("demographics")
check_members("housing", FALSE)
}
