% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outreach.R
\name{check_outreach}
\alias{check_outreach}
\title{Check Outreach}
\usage{
check_outreach(check = "member")
}
\arguments{
\item{check}{The check to be performed.  Choices are member, date, type, or outcome}
}
\value{
a flextable with the results of the requested check
}
\description{
Check Outreach
}
\examples{
check_outreach()
check_outreach("type")
}
