% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_progress_notes.R
\name{check_progress_notes}
\alias{check_progress_notes}
\title{Check Progress Notes}
\usage{
check_progress_notes(check = "staff")
}
\arguments{
\item{check}{The check to be performed.  One of staff, member, goal, text}
}
\value{
a flextable containing the requested checks
}
\description{
Check Progress Notes
}
\examples{
check_progress_notes()
check_progress_notes("staff")
}
