% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_supports.R
\name{check_supports}
\alias{check_supports}
\title{Check Supports}
\usage{
check_supports(check = "staff")
}
\arguments{
\item{check}{The name of a check to be performed.  One of staff, member, date, duration, type, durations}
}
\value{
a flextable with the requested checks
}
\description{
Check Supports
}
\examples{
check_supports()
check_supports("duration") # list supports with no duration recorded
check_supports("durations") # drequency table of durations
}
