% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_employment.R
\name{current_employment}
\alias{current_employment}
\title{Current Employment}
\usage{
current_employment()
}
\value{
a flextable showing the current employment of members
}
\description{
Must have support as one of IE, TE, SE, and no end date
}
\examples{
current_employment()
}
