% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_city_county.R
\name{dict_city_county}
\alias{dict_city_county}
\title{dict_city_county}
\usage{
dict_city_county()
}
\value{
dictionary of counties for cities.  key is city name, value is county name
}
\description{
dict_city_county
}
\examples{
dict_city_county()
}
