% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_support_types.R
\name{dict_support_types}
\alias{dict_support_types}
\title{dict_support_types}
\usage{
dict_support_types()
}
\value{
Dictionary of support types.  key, value, and "narrow" -- a collection of values optimized for narrow plots and lists
}
\description{
dict_support_types
}
\examples{
dict_support_types()
}
