% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_table_names.R
\name{dict_table_names}
\alias{dict_table_names}
\title{Dictionary of table names}
\usage{
dict_table_names()
}
\value{
a list with two elements -- \code{key}, a vector of keys, and \code{value}, a vector of values
}
\description{
Provides a translation between the external name of the table, known as the value: contacts, members, etc, and the internal name of the table, known as the key.  In this way, a dict_ function performs like a python dict object.
}
\examples{
dict_table_names()$key[2] # key value for contacts table
dict_table_names()$key[which(dict_table_names()$value == "contacts")]
}
