% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flourish.connection.R
\name{flourish.connection}
\alias{flourish.connection}
\title{Establish a database connection to Flourish}
\usage{
flourish.connection()
}
\value{
a database connection
}
\description{
Establish a database connection to Flourish
}
\examples{
con <- flourish.connection()
RMySQL::dbDisconnect(con)


}
