% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flourish_connection.R
\name{flourish_connection}
\alias{flourish_connection}
\title{Establish a database connection to Flourish}
\usage{
flourish_connection()
}
\value{
a database connection
}
\description{
Establish a database connection to Flourish
}
\examples{
con <- flourish_connection()
RMySQL::dbDisconnect(con)
}
