% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.attendance.R
\name{get.attendance}
\alias{get.attendance}
\title{Get attendance data from Flourish}
\usage{
get.attendance(include.deleted = FALSE, days = Inf)
}
\arguments{
\item{include.deleted}{if TRUE, include deleted attendance records, default is FALSE}

\item{days}{Number of days from today to include.  Default is Infinite, include all days.  A common value is 90, include all attendance records for the previous 90 days.  Members with attendance in the past 90 days are defined as active members.}
}
\value{
data.frame of attendance records
}
\description{
Get attendance data from Flourish
}
\examples{
nrow(get.attendance()) # number of non-deleted attendance records all time
get.attendance() # get all non-deleted attendance records as a data.frame
get.attendance(days=90) # attendance over the previous 90 days from today.
}
