% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.average.daily.attendance.R
\name{get.average.daily.attendance}
\alias{get.average.daily.attendance}
\title{Get average daily attendance}
\usage{
get.average.daily.attendance(days = 30)
}
\arguments{
\item{days}{The number of days from today to compute average daily attendance. days=7 is the past week.}
}
\value{
a number -- the average daily attendance for the number of days specified
}
\description{
Average daily attendance is the average of daily attendance for non-weekend days in a period of time.
}
\examples{
get.average.daily.attendance() # last 30 days
get.average.daily.attendance(180) # previous half year
}
