% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.campaign.names.R
\name{get.campaign.names}
\alias{get.campaign.names}
\title{Get the campaign names}
\usage{
get.campaign.names()
}
\value{
a list with two elements -- \code{key}, a vector of keys, and \code{value}, a vector of values
}
\description{
Get the campaign names
}
\examples{
get.campaign.names()$key[2] # key value for 2018 Amazing Give
}
