% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.employment.R
\name{get.employment}
\alias{get.employment}
\title{Get employment data from Flourish}
\usage{
get.employment(include.deleted = FALSE)
}
\arguments{
\item{include.deleted}{if TRUE, include deleted employment records}
}
\value{
a data.frame of employment records
}
\description{
Get employment data from Flourish
}
\examples{
emp <- get.employment()
goc.table(emp, "Employment", paste0(nrow(emp), " records as of ", Sys.Date()))
}
