% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.gifts.R
\name{get.gifts}
\alias{get.gifts}
\title{Get gifts from Flourish}
\usage{
get.gifts(include.deleted = FALSE)
}
\arguments{
\item{include.deleted}{If TRUE, include deleted gifts}
}
\value{
a data.frame of gifts
}
\description{
Get gifts from Flourish
}
\examples{
get.gifts() # All non-deleted gifts
hist(get.gifts()[,c("flo_gift_amount")]) # Histogram of all gift amounts
}
