% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.goals.R
\name{get.goals}
\alias{get.goals}
\title{Get goal records from Flourish}
\usage{
get.goals(include.deleted = FALSE)
}
\arguments{
\item{include.deleted}{If TRUE, include deleted records}
}
\value{
a data.fram of goal records
}
\description{
Get goal records from Flourish
}
\examples{
get.goals()
}
