% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.holidays.R
\name{get.holidays}
\alias{get.holidays}
\title{Get GOC holidays 2021-24}
\usage{
get.holidays()
}
\value{
vector of GOC holidays each in the format "yyy-mm-dd"
}
\description{
Get GOC holidays 2021-24
}
\examples{
get.holidays()[14] # Thanksgiving Day, 2022 is 2022-11-24
}
