% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.outreach.R
\name{get.outreach}
\alias{get.outreach}
\title{Get outreach records from Flourish}
\usage{
get.outreach(include.deleted = FALSE)
}
\arguments{
\item{include.deleted}{If TRUE, included deleted records}
}
\value{
a data.frame of outreach records
}
\description{
Get outreach records from Flourish
}
\examples{
get.outreach()
}
