% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.progress.notes.R
\name{get.progress.notes}
\alias{get.progress.notes}
\title{Get progress notes from Flourish}
\usage{
get.progress.notes(include.deleted = FALSE)
}
\arguments{
\item{include.deleted}{if TRUE, include deleted progress notes}
}
\value{
a data fram with progress.notes
}
\description{
Get progress notes from Flourish
}
\examples{
get.progress.notes() # get non-deleted progress notes
get.progress.notes(include.deleted=TRUE) # get all progress notes, including deleted
}
