% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.staff.names.R
\name{get.staff.names}
\alias{get.staff.names}
\title{Get names of staff members}
\usage{
get.staff.names()
}
\value{
a vector of staff names.  Each name is "last, first"
}
\description{
Get names of staff members
}
\examples{
length(get.staff.names()) # count number of staff
}
