% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.table.names.R
\name{get.table.names}
\alias{get.table.names}
\title{Get a vector containing the names of the tables in a database}
\usage{
get.table.names()
}
\value{
a vector of table names
}
\description{
Get a vector containing the names of the tables in a database
}
\examples{
names <- get.table.names()


}
