% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.users.R
\name{get.users}
\alias{get.users}
\title{Get users table from Flourish}
\usage{
get.users(include.deleted = FALSE)
}
\arguments{
\item{include.deleted}{if TRUE, include deleted records.  Default is FALSE}
}
\value{
a data.frame corresponding to the the SQL users table in Flourish
}
\description{
Get users table from Flourish
}
\examples{
get.users() # return all non-deleted records
get.users(include.deleted=TRUE) # return all records, including deleted
get.users()[,c("id","date_entered")] # return named columns for non-deleted records
colnames(get.users()) # return the column names of the users table
}
