% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_average_daily_attendance.R
\name{get_average_daily_attendance}
\alias{get_average_daily_attendance}
\title{Get average daily attendance}
\usage{
get_average_daily_attendance(days = 30)
}
\arguments{
\item{days}{The number of days from today to compute average daily attendance. days=7 is the past week.}
}
\value{
a number -- the average daily attendance for the number of days specified
}
\description{
Average daily attendance is the average of daily attendance for non-weekend days in a period of time.
}
\examples{
get_average_daily_attendance() # last 30 days
get_average_daily_attendance(180) # previous half year
}
