% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_birthdays.R
\name{get_birthdays}
\alias{get_birthdays}
\title{Get member birthdays for a date}
\usage{
get_birthdays(date = Sys.Date())
}
\arguments{
\item{date}{Members with birthdays on this day will be found}
}
\value{
Return a text string of first names of members whose birthday falls on the date.  If none, return the string "No birthdays today"
}
\description{
Only active members are included.  Deceased members are not included.  On leave members are included.
}
\examples{
get_birthdays() # get birthdays for today
get_birthdays("2022-05-01")
}
