% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_campaign_names.R
\name{get_campaign_names}
\alias{get_campaign_names}
\title{Get the campaign names}
\usage{
get_campaign_names()
}
\value{
a list with two elements -- \code{key}, a vector of keys, and \code{value}, a vector of values
}
\description{
Get the campaign names
}
\examples{
get_campaign_names()$key[2] # key value for 2018 Amazing Give
}
