% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contacts.R
\name{get_contacts}
\alias{get_contacts}
\title{Get contact data from Flourish}
\usage{
get_contacts(include.deleted = FALSE, include.email = T)
}
\arguments{
\item{include.deleted}{if TRUE, deleted contacts are included}

\item{include.email}{if TRUE, the primary emaiil address for the contact is included in the data returned from Flourish}
}
\value{
a data.frame containing contact data
}
\description{
Get contact data from Flourish
}
\examples{
get_contacts()
contacts <- get_contacts(include.email = TRUE)
nrow(contacts[!is.na(contacts$email_address), ]) # count records with email
}
