% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_employment.R
\name{get_employment}
\alias{get_employment}
\title{Get employment data from Flourish}
\usage{
get_employment(include.deleted = FALSE)
}
\arguments{
\item{include.deleted}{if TRUE, include deleted employment records}
}
\value{
a data.frame of employment records
}
\description{
Get employment data from Flourish
}
\examples{
emp <- get_employment()
goc_table(emp, "Employment", paste0(nrow(emp), " records as of ", Sys.Date()))
}
