% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_field_notes.R
\name{get_field_notes}
\alias{get_field_notes}
\title{Get Field Notes}
\usage{
get_field_notes(table = "attendance")
}
\arguments{
\item{table}{Field notes are returned for the table named.  Tables are named using common names, not internal names -- attendance, contacts, etc.}
}
\value{
a data.frame containing the field notes for the desired table. The data.frame has one row for every field in the desired table.  The data.frame has four columns -- field_name, auto, action, notes
}
\description{
Get Field Notes
}
\examples{
get_field_notes("members") # notes on the fields in the member table
}
