% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gifts.R
\name{get_gifts}
\alias{get_gifts}
\title{Get gifts from Flourish}
\usage{
get_gifts(include.deleted = FALSE)
}
\arguments{
\item{include.deleted}{If TRUE, include deleted gifts}
}
\value{
a data.frame of gifts
}
\description{
Get gifts from Flourish
}
\examples{
get_gifts() # All non-deleted gifts
hist(get_gifts()[,c("flo_gift_amount")]) # Histogram of all gift amounts
}
