% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_goals.R
\name{get_goals}
\alias{get_goals}
\title{Get goal records from Flourish}
\usage{
get_goals(include.deleted = FALSE)
}
\arguments{
\item{include.deleted}{If TRUE, include deleted records}
}
\value{
a data.fram of goal records
}
\description{
Get goal records from Flourish
}
\examples{
get_goals()
}
