% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_holidays.R
\name{get_holidays}
\alias{get_holidays}
\title{Get GOC holidays 2021-24}
\usage{
get_holidays()
}
\value{
vector of GOC holidays each in the format "yyy-mm-dd"
}
\description{
Get GOC holidays 2021-24
}
\examples{
get_holidays()[14] # Thanksgiving Day, 2022 is 2022-11-24
}
