% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_members.R
\name{get_members}
\alias{get_members}
\title{Get member data from Flourish}
\usage{
get_members(include.deleted = FALSE, active.only = FALSE)
}
\arguments{
\item{include.deleted}{if TRUE, include deleted records.  Default is FALSE}

\item{active.only}{if TRUE, include only those members who have attended in the past 90 days, are not
deceased and are not on leave.}
}
\value{
a data.frame of member data
}
\description{
Get member data from Flourish
}
\examples{
get_members()
get_members(active.only=TRUE)
}
