% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outreach.R
\name{get_outreach}
\alias{get_outreach}
\title{Get outreach records from Flourish}
\usage{
get_outreach(include.deleted = FALSE)
}
\arguments{
\item{include.deleted}{If TRUE, included deleted records}
}
\value{
a data.frame of outreach records
}
\description{
Get outreach records from Flourish
}
\examples{
get_outreach()
}
