% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_progress_notes.R
\name{get_progress_notes}
\alias{get_progress_notes}
\title{Get progress notes from Flourish}
\usage{
get_progress_notes(include.deleted = FALSE)
}
\arguments{
\item{include.deleted}{if TRUE, include deleted progress notes}
}
\value{
a data frame with progress.notes
}
\description{
Get progress notes from Flourish
}
\examples{
get_progress_notes() # get non-deleted progress notes
get_progress_notes(include.deleted=TRUE) # get all progress notes, including deleted
}
