% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_staff_names.R
\name{get_staff_names}
\alias{get_staff_names}
\title{Get names of staff members}
\usage{
get_staff_names()
}
\value{
a vector of staff names.  Each name is "last, first"
}
\description{
Get names of staff members
}
\examples{
length(get_staff_names()) # count number of staff
}
