% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_supports.R
\name{get_supports}
\alias{get_supports}
\title{Get supports from Flourish}
\usage{
get_supports(
  include.deleted = FALSE,
  with.members = FALSE,
  active.only = FALSE
)
}
\arguments{
\item{include.deleted}{if TRUE, include deleted supports}

\item{with.members}{if TRUE, return outer join of support data with member data.  WARNING: the result can be a large data.frame}

\item{active.only}{if TRUE, and with.members=TRUE, the outer join of supports and members include only active members}
}
\value{
a data frame of supports
}
\description{
Get supports from Flourish
}
\examples{
get_supports()
get_supports(with.members=TRUE) # return outer join of members and supports (large data.frame)
}
