% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{Return the contents of a Flourish table as a data.frame}
\usage{
get_table(table.name, include.deleted = FALSE, where = "")
}
\arguments{
\item{table.name}{Name of a table in Flourish}

\item{include.deleted}{If True, return deleted records as well as well as non-deleted}

\item{where}{A SQL where clause to be inserted into the database query.  Include the word WHERE}
}
\value{
A data.frame containing the requested table
}
\description{
Return the contents of a Flourish table as a data.frame
}
\examples{
get_table("sa_employment") # return the employment table as a data.frame

}
\seealso{
\code{\link[=get_table_names]{get_table_names()}} to get a list of table names in a Flourish database
}
