% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_names.R
\name{get_table_names}
\alias{get_table_names}
\title{Get a vector containing the names of the tables in a database}
\usage{
get_table_names()
}
\value{
a vector of table names
}
\description{
Get a vector containing the names of the tables in a database
}
\examples{
names <- get_table_names()
}
