% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users.R
\name{get_users}
\alias{get_users}
\title{Get users table from Flourish}
\usage{
get_users(include.deleted = FALSE)
}
\arguments{
\item{include.deleted}{if TRUE, include deleted records.  Default is FALSE}
}
\value{
a data.frame corresponding to the the SQL users table in Flourish
}
\description{
Get users table from Flourish
}
\examples{
get_users() # return all non-deleted records
get_users(include.deleted = TRUE) # return all records, including deleted
get_users()[, c("id", "date_entered")] # return named columns for non-deleted records
colnames(get_users()) # return the column names of the users table
}
