% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goc.table.R
\name{goc.table}
\alias{goc.table}
\title{Prettify display of a data.frame
Use \code{set_flextable_defaults} to control look and feel}
\usage{
goc.table(tab, caption = "")
}
\arguments{
\item{tab}{the data.frame to be prettified}

\item{caption}{an optional caption to display at the top of the display}
}
\value{
a flextable
}
\description{
Prettify display of a data.frame
Use \code{set_flextable_defaults} to control look and feel
}
\examples{
goc.table(get.table('users'), 'User accounts on GOC Flourish')
}
