% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goc.theme.R
\name{goc.theme}
\alias{goc.theme}
\title{GOC theme for flextables}
\usage{
goc.theme(x)
}
\arguments{
\item{x}{The table to get styles}
}
\value{
x the table that was entered, with styling
}
\description{
Used in \code{goc.table} to produce attractive, standard tables
}
\examples{
goc.theme(flextable::flextable(data.frame(x=seq(1,5),y=runif(5))))

}
