% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goc_plot.R
\name{goc_plot}
\alias{goc_plot}
\title{GOC Plot}
\usage{
goc_plot(p)
}
\arguments{
\item{p}{a plot to be goc-styled}
}
\value{
a plot that is goc-styled
}
\description{
GOC Plot
}
\examples{
d <- data.frame(x=1:5, y=4:8)
goc_plot(ggplot2::ggplot(d, ggplot2::aes(x,y)) + ggplot2::geom_line())
}
