% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goc_table.R
\name{goc_table}
\alias{goc_table}
\title{Prettify display of a data.frame}
\usage{
goc_table(df, heading = "", caption = "")
}
\arguments{
\item{df}{the data.frame to be prettified}

\item{heading}{an optional heading to display at the top of the display}

\item{caption}{an optional caption to display at the top of the display}
}
\value{
a prettified flextable
}
\description{
Given a data.frame and an optional caption, display the data.frame as a table suitable for printing, web display, or saving. Use \code{set_flextable_defaults} to control look and feel
}
\examples{
goc_table(get_table("users"), "User accounts on GOC Flourish")
}
