% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goc_theme.R
\name{goc_theme}
\alias{goc_theme}
\title{GOC theme for flextables}
\usage{
goc_theme(x)
}
\arguments{
\item{x}{The table to get styles}
}
\value{
x the table that was entered, with styling
}
\description{
Used in \code{goc_table} to produce attractive, standard tables
}
\examples{
goc_theme(flextable::flextable(data.frame(x=seq(1,5),y=runif(5))))
}
