% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_accessibility.R
\name{list_accessibility}
\alias{list_accessibility}
\title{List Accessibility Restrictions of Members}
\usage{
list_accessibility()
}
\value{
a flextable report of member accessibility restrictions
}
\description{
List Accessibility Restrictions of Members
}
\examples{
list_accessibility()
}
