% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_active_member_outreach.R
\name{list_active_member_outreach}
\alias{list_active_member_outreach}
\title{Active Member Outreach List}
\usage{
list_active_member_outreach()
}
\value{
a flextable showing active members with phone and email who haven't been seen in more than a day
}
\description{
Active Member Outreach List
}
\examples{
list_active_member_outreach()
}
