% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_all_donors.R
\name{list_all_donors}
\alias{list_all_donors}
\title{List All Donors}
\usage{
list_all_donors()
}
\value{
a flex table listing all donors and total dollars donated
}
\description{
List All Donors
}
\examples{
list_all_donors()
}
