% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_all_member_outreach.R
\name{list_all_member_outreach}
\alias{list_all_member_outreach}
\title{All Member Outreach List}
\usage{
list_all_member_outreach()
}
\value{
a flextable showing all members with phone and email who haven't been seen in more than a day
}
\description{
All Member Outreach List
}
\examples{
list_all_member_outreach()
}
