% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_allergies.R
\name{list_allergies}
\alias{list_allergies}
\title{List Member Allergies}
\usage{
list_allergies()
}
\value{
a flextable list of members and their allergies
}
\description{
List Member Allergies
}
\examples{
list_allergies()
}
