% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_attendance_checklist.R
\name{list_attendance_checklist}
\alias{list_attendance_checklist}
\title{Attendance Checklist}
\usage{
list_attendance_checklist(date = Sys.Date())
}
\arguments{
\item{date}{Date for the attendance checklist}
}
\value{
a flextable with four columns -- name, time in, time out, and hours.  Blank indicates no value.
}
\description{
For a given date, show attendance for that date -- name, time in, time out, and hours. Include
staff whether there is an attendance record or not.
}
\examples{
list_attendance_checklist()
list_attendance_checklist("2023-05-04")
}
