% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_birthdays.R
\name{list_birthdays}
\alias{list_birthdays}
\title{List Birthdays}
\usage{
list_birthdays(date = Sys.Date())
}
\arguments{
\item{date}{The month to list the birthdays specified by any day within the month.}
}
\value{
a flextable listing the members with birthdays during the month, ordered by birthday
}
\description{
Lists member birthdays
}
\examples{
list_birthdays()
}
