% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_campaign_donors.R
\name{list_campaign_donors}
\alias{list_campaign_donors}
\title{List Campaign donors}
\usage{
list_campaign_donors(gift_campaign = "2022 Amazing Give")
}
\arguments{
\item{gift_campaign}{value for campaign, for example "2018 Amazing Give"}
}
\value{
a flextable showing donors to the selected campaign
}
\description{
Mutliple gifts by a donor to a campaign are totaled
}
\examples{
list_campaign_donors()
list_campaign_donors("2022 Annual Campaign")
}
