% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_clubhouse_billing.R
\name{list_clubhouse_billing}
\alias{list_clubhouse_billing}
\title{Clubhouse Billing List}
\usage{
list_clubhouse_billing(date = Sys.Date())
}
\arguments{
\item{date}{The date on which the billing list will end.  The start date is always the first of the
month of the month that contains the end date.  So, for example, if the date is set to 2023-05-18,
the start date is 2023-05-01 and the end date is 2023-05-18.}
}
\value{
a flextable of billing line items suitable for entry into the KIS system required by LSF.
}
\description{
Create a flextable that shows the billing line items needed for Lutheran Services (LSF) KIS (Knight
Information Systems) billing. The report processes attendance data for staff and members and
creates billable time records for each staff member.
}
\examples{
list_clubhouse_billing()
list_clubhouse_billing("2023-05-13") # Through the first two weeks of the month
}
