% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_contacts.R
\name{list_contacts}
\alias{list_contacts}
\title{List Contacts}
\usage{
list_contacts(type = "families")
}
\arguments{
\item{type}{One of "families" "people" "organizations"}
}
\value{
a flextable with selected contacts
}
\description{
List Contacts
}
\examples{
list_contacts()
}
