% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_current_employment.R
\name{list_current_employment}
\alias{list_current_employment}
\title{Current Employment}
\usage{
list_current_employment()
}
\value{
a flextable showing the current employment of members
}
\description{
Must have support as one of IE, TE, SE, and no end date
}
\examples{
list_current_employment()
}
