% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_fields.R
\name{list_fields}
\alias{list_fields}
\title{List Fields in a Table}
\usage{
list_fields(table_name = "attendance")
}
\arguments{
\item{table_name}{One of the Flourish table names. Tables can be referred to by common names, eg, contacts, members, etc, or by database names such as flo_gifts}
}
\value{
a flextable showing the fields in the table, the number of times the field is used, whether the field is auto-populated, notes, and actions
}
\description{
List Fields in a Table
}
\examples{
list_fields()
list_fields("attendance") # keyword call
list_fields("email_addr_bean_rel") # direct call by name
}
