% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_gift_amounts_by_campaign.R
\name{list_gift_amounts_by_campaign}
\alias{list_gift_amounts_by_campaign}
\title{Gift Amounts by Campaign}
\usage{
list_gift_amounts_by_campaign()
}
\value{
a flextable with rows for each campaign, and columns for gift sizes
}
\description{
Gift Amounts by Campaign
}
\examples{
list_gift_amounts_by_campaign()
}
