% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_gift_statement.R
\name{list_gift_statement}
\alias{list_gift_statement}
\title{List Gift Statement}
\usage{
list_gift_statement(contact_id, year)
}
\arguments{
\item{contact_id}{The Contact ID of the contact whose statement is need}

\item{year}{The year for which the statement is needed}
}
\value{
a flextable containing the statement
}
\description{
List Gift Statement
}
\examples{
list_gift_statement("Contact-03914", 2022)
list_gift_statement("Contact-04044", 2022)
}
