% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_gifts.R
\name{list_gifts}
\alias{list_gifts}
\title{List Gifts}
\usage{
list_gifts(date = Sys.Date())
}
\arguments{
\item{date}{Gifts will be listed inclusive of the date, back two years}
}
\value{
a flextable containing the gifts
}
\description{
List Gifts
}
\examples{
list_gifts() # list gifts in the last year from today
list_gifts("2023-06-30") # list gifts in the last fiscal year
list_gifts("2022-12-31") # list gifts in the last calendar year
}
