% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_gifts_by_campaign.R
\name{list_gifts_by_campaign}
\alias{list_gifts_by_campaign}
\title{List Gifts by Campaign}
\usage{
list_gifts_by_campaign()
}
\value{
a flextable listing the number of gifts and donors for each campaign
}
\description{
List Gifts by Campaign
}
\examples{
list_gifts_by_campaign()
}
