% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_gifts_by_payment_method_and_time.R
\name{list_gifts_by_payment_method_and_time}
\alias{list_gifts_by_payment_method_and_time}
\title{List Gifts by Payment Method and Time}
\usage{
list_gifts_by_payment_method_and_time(by.year = T)
}
\arguments{
\item{by.year}{If TRUE, table is one row per year.  If FALSE, result is by year/month}
}
\value{
a flextable with gifts counted by payment type and time (year or year/month)
}
\description{
List Gifts by Payment Method and Time
}
\examples{
list_gifts_by_payment_method_and_time()
}
