% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_goal_types_by_staff.R
\name{list_goal_types_by_staff}
\alias{list_goal_types_by_staff}
\title{List Goal Types by Staff}
\usage{
list_goal_types_by_staff(date = Sys.Date(), unit = "all")
}
\arguments{
\item{date}{Date around which the unit will be calculated}

\item{unit}{Time period to be displayed from week, month, year, all}
}
\value{
a flextable showing goal types by staff for the desired time from the request date
}
\description{
List Goal Types by Staff
}
\examples{
list_goal_types_by_staff()
list_goal_types_by_staff(date="2023-01-01", unit="week")
}
