% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_goals_by_member_and_staff.R
\name{list_goals_by_member_and_staff}
\alias{list_goals_by_member_and_staff}
\title{Goals by Active Member and Staff}
\usage{
list_goals_by_member_and_staff()
}
\value{
a flextable showing goals for each member by staff
}
\description{
Goals by Active Member and Staff
}
\examples{
list_goals_by_member_and_staff()
}
