% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_last_attendance.R
\name{list_last_attendance}
\alias{list_last_attendance}
\title{List Last Attendnce in a Month}
\usage{
list_last_attendance(date = Sys.Date())
}
\arguments{
\item{date}{The month of interest, specified by any day in the month.  For example: "2023-03-20"
produces a list of the last attendance of members in the month March, 2023}
}
\value{
a flextable listing the date of last attendance for each member in the speciefied month
}
\description{
List Last Attendnce in a Month
}
\examples{
list_last_attendance()
list_last_attendance("2023-01-03")
}
