% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_member_signin.R
\name{list_member_signin}
\alias{list_member_signin}
\title{List Member Sign-In Sheet}
\usage{
list_member_signin(date = Sys.Date())
}
\arguments{
\item{date}{The date for which sign-in is needed}
}
\value{
a flextable with the sign-in sheet
}
\description{
Active Members are listed along with staff and volunteers who have attended in the last 90 days
}
\examples{
list_member_signin()
list_member_signin("2023-05-21")
}
