% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_members_by_city.R
\name{list_members_by_city}
\alias{list_members_by_city}
\title{List Members By City}
\usage{
list_members_by_city()
}
\value{
a flextable with one row per city, including count of members and county of city
}
\description{
List Members By City
}
\examples{
list_members_by_city()
}
