% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_members_by_county.R
\name{list_members_by_county}
\alias{list_members_by_county}
\title{List Members By County}
\usage{
list_members_by_county()
}
\value{
a flextable with one row per county, including count of members
}
\description{
List Members By County
}
\examples{
list_members_by_county()
}
