% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_members_by_year.R
\name{list_members_by_year}
\alias{list_members_by_year}
\title{Members By Year}
\usage{
list_members_by_year(output = "flextable")
}
\arguments{
\item{output}{If "flextable" a flextable is returned with one row per year and the number of members attending in that year counted.  Otherwise a data.frame is returned.}
}
\value{
either a flextable or a data.frame, depending on the value of the output parameter.  One row per year, count of the number of members attending at least once in that year.
}
\description{
Members By Year
}
\examples{
list_members_by_year()
}
