% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_monthly_attendance_summary.R
\name{list_monthly_attendance_summary}
\alias{list_monthly_attendance_summary}
\title{Monthly Attendance Summary}
\usage{
list_monthly_attendance_summary(date = Sys.Date())
}
\arguments{
\item{date}{Date to include the month to be displayed}
}
\value{
a flextable showing for each member, the number of hours attended
}
\description{
Monthly Attendance Summary
}
\examples{
list_monthly_attendance_summary()
list_monthly_attendance_summary("2022-03-05")
}
