% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_monthly_progress_notes.R
\name{list_monthly_progress_notes}
\alias{list_monthly_progress_notes}
\title{List Monthly Progress Notes}
\usage{
list_monthly_progress_notes(date = Sys.Date())
}
\arguments{
\item{date}{date in month for desired report}
}
\value{
a flextable of members who attended during the month, with their progress notes
}
\description{
List Monthly Progress Notes
}
\examples{
list_monthly_progress_notes()
list_monthly_progress_notes("2203-03-05")
}
