% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_monthly_supports_by_member.R
\name{list_monthly_supports_by_member}
\alias{list_monthly_supports_by_member}
\title{Monthly Supports by Member}
\usage{
list_monthly_supports_by_member(date = Sys.Date())
}
\arguments{
\item{date}{A date in the month the table will be made.  For example if date = "2023-04-12", the table will be made for April, 2023.}
}
\value{
a flextable with one row per member attending in the selected month, columns for support types and cells with counts of the support type for each member
}
\description{
Monthly Supports by Member
}
\examples{
list_monthly_supports_by_member()
list_monthly_supports_by_member("2023-04-12")
}
