% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_new_members.R
\name{list_new_members}
\alias{list_new_members}
\title{List New Members}
\usage{
list_new_members(date = Sys.Date(), ndays = 10)
}
\arguments{
\item{date}{The date of the report}

\item{ndays}{The number of days prior to the report date that a member will be considered new.}
}
\value{
a flextable with the names of the new members and their membership dates
}
\description{
List New Members
}
\examples{
list_new_members()
list_new_members("2023-05-22", ndays = 30)
}
