% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_newly_absent_members.R
\name{list_newly_absent_members}
\alias{list_newly_absent_members}
\title{List Newly Absent Members}
\usage{
list_newly_absent_members(date = Sys.Date())
}
\arguments{
\item{date}{Date to be checked for newly absent members}
}
\value{
a flextable showing members absent more than 90 days and less than 180 days
}
\description{
List Newly Absent Members
}
\examples{
list_newly_absent_members()
}
