% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_organizations.R
\name{list_organizations}
\alias{list_organizations}
\title{List Organization Contact Information}
\usage{
list_organizations()
}
\value{
a flextable with all organizations and their contact information
}
\description{
List Organization Contact Information
}
\examples{
list_organizations()
}
