% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_people.R
\name{list_people}
\alias{list_people}
\title{List People Contact Information}
\usage{
list_people()
}
\value{
a flextable with all people and their contact information
}
\description{
List People Contact Information
}
\examples{
list_people()
}
