% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_record_counts.R
\name{list_record_counts}
\alias{list_record_counts}
\title{List record counts}
\usage{
list_record_counts()
}
\value{
a flextable of record counts for each table
}
\description{
List record counts
}
\examples{
list_record_counts()
}
