% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_referral_summary.R
\name{list_referral_summary}
\alias{list_referral_summary}
\title{List Referring Agencies}
\usage{
list_referral_summary()
}
\value{
a flextable of agencies, the number of referrals they have made, and the date of last referral
}
\description{
List Referring Agencies
}
\examples{
list_referral_summary()
}
