% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_staff_progress_notes.R
\name{list_staff_progress_notes}
\alias{list_staff_progress_notes}
\title{Staff Progress Notes}
\usage{
list_staff_progress_notes(date = Sys.Date(), interval = "month")
}
\arguments{
\item{date}{A date in the interval of interest}

\item{interval}{One of "week", "month" year", "all"}
}
\value{
a flex table of progress notes by staff member in the interval requested
}
\description{
Staff Progress Notes
}
\examples{
list_staff_progress_notes()
list_staff_progress_notes(date = "2023-04-05", interval = "month") # progress notes for April
}
