% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_staff_time.R
\name{list_staff_time}
\alias{list_staff_time}
\title{List Staff Time}
\usage{
list_staff_time(name, start_date)
}
\arguments{
\item{name}{Name of an employee in the form last, first}

\item{start_date}{start date of the period to be reported}
}
\value{
a flextable showing name, date, time in and time out by rows
}
\description{
List Staff Time
}
\examples{
list_staff_time(name = "Reynolds, Keisha", start_date = "2022-01-01") # Fictional name
}
