% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_supports_by_staff_and_type.R
\name{list_supports_by_staff_and_type}
\alias{list_supports_by_staff_and_type}
\title{List Supports by Staff and Type}
\usage{
list_supports_by_staff_and_type()
}
\value{
a flextable with rows for each member and columns for each support type
}
\description{
List Supports by Staff and Type
}
\examples{
list_supports_by_staff_and_type()
}
