% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_supports_by_year.R
\name{list_supports_by_year}
\alias{list_supports_by_year}
\title{Supports by Year}
\usage{
list_supports_by_year()
}
\value{
a flextable showing support types (rows) and years (columns).  Entries are the counts of supports of that type for the year.
}
\description{
Supports by Year
}
\examples{
list_supports_by_year()
}
