% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tracking_by_member_and_staff.R
\name{list_tracking_by_member_and_staff}
\alias{list_tracking_by_member_and_staff}
\title{Tracking Data by Member and Staff}
\usage{
list_tracking_by_member_and_staff()
}
\value{
a flextable showing active members, their assigned staff, and tracking data consisting of number of visits in the
past 90 days and date of most recent visit, supports and most recent support, open goals and date of most recent goal, and
progress notes in the last 90 days and date of most recent progress note
}
\description{
Tracking Data by Member and Staff
}
\examples{
list_tracking_by_member_and_staff()
}
