% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_visitors_sign_in.R
\name{list_visitors_sign_in}
\alias{list_visitors_sign_in}
\title{Visitor's Sign In Sheet}
\usage{
list_visitors_sign_in(date = Sys.Date())
}
\arguments{
\item{date}{The date for the sign in sheet}
}
\value{
a flextable with the visitor's sign in sheet for the requested date
}
\description{
Visitor's Sign In Sheet
}
\examples{
list_visitors_sign_in()
list_visitors_sign_in("2015-07-04")
}
