% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.intersection.R
\name{make.intersection}
\alias{make.intersection}
\title{Make intersection of two intervals}
\usage{
make.intersection(int1, int2)
}
\arguments{
\item{int1}{an interval}

\item{int2}{a second interval}
}
\value{
an interval which is the intersection of the two intervals.  If the intervals do not intersect, return NA
}
\description{
Make intersection of two intervals
}
\examples{
int1 <- lubridate::interval(lubridate::ymd("2023-04-01"), lubridate::ymd("2023-04-10"))
int2 <- lubridate::interval(lubridate::ymd("2023-04-03"), lubridate::ymd("2023-04-24"))
int3 <- make.intersection(int1, int2) # from 3 to 10
}
