% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_active_members_over_time.R
\name{plot_active_members_over_time}
\alias{plot_active_members_over_time}
\title{Plot Active Members over Time}
\usage{
plot_active_members_over_time(start_date = Sys.Date() - lubridate::years(1))
}
\arguments{
\item{start_date}{plot runs from start date to today}
}
\value{
a plot showing active members over time
}
\description{
Plot Active Members over Time
}
\examples{
plot_active_members_over_time()
}
