% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_attendance_by_year.R
\name{plot_attendance_by_year}
\alias{plot_attendance_by_year}
\title{Plot Attendance by Year}
\usage{
plot_attendance_by_year()
}
\value{
plot of attendance by year with lines for staff, members, and volunteers
}
\description{
Plot Attendance by Year
}
\examples{
plot_attendance_by_year()
}
