% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_daily_supports_by_type.R
\name{plot_daily_supports_by_type}
\alias{plot_daily_supports_by_type}
\title{Plot Daily Supports by Type}
\usage{
plot_daily_supports_by_type(date = Sys.Date())
}
\arguments{
\item{date}{date of plot}
}
\value{
a plot showing vertical bars with counts of each support type on the day specified
}
\description{
Plot Daily Supports by Type
}
\examples{
plot_daily_supports_by_type() # often blank -- supports have not been entered for the current day
plot_daily_supports_by_type("2023-03-09")
}
