% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gifts_by_year.R
\name{plot_gifts_by_year}
\alias{plot_gifts_by_year}
\title{Plot gifts by calendar year}
\usage{
plot_gifts_by_year(date = Sys.Date())
}
\arguments{
\item{date}{Show gifts up to to the end of the year preceding the date}
}
\value{
a plot showing total raised each year in vertical bars
}
\description{
Plot gifts by calendar year
}
\examples{
plot_gifts_by_year()
}
