% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gifts_vs_total_raised.R
\name{plot_gifts_vs_total_raised}
\alias{plot_gifts_vs_total_raised}
\title{Plot Number of Gifts vs Total Raised for each campaign}
\usage{
plot_gifts_vs_total_raised()
}
\value{
dot plot of gifts vs raised, labeled by campaign
}
\description{
Note: small campaigns are excluded
}
\examples{
plot_gifts_vs_total_raised()
}
