% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_member_attendance_by_month.R
\name{plot_member_attendance_by_month}
\alias{plot_member_attendance_by_month}
\title{Plot Member attendance by Month}
\usage{
plot_member_attendance_by_month(
  start_date = as_date("2008-01-01"),
  end_date = Sys.Date()
)
}
\arguments{
\item{start_date}{The start date of the plot}

\item{end_date}{The end date for the plot}
}
\value{
a line plot of attendance from the start date to the end date
}
\description{
Plot Member attendance by Month
}
\examples{
plot_member_attendance_by_month()
plot_member_attendance_by_month(lubridate::as_date("2020-01-01"))
}
