% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_membership_by_the_numbers.R
\name{plot_membership_by_the_numbers}
\alias{plot_membership_by_the_numbers}
\title{Plot Membership by the Numbers}
\usage{
plot_membership_by_the_numbers(date = Sys.Date(), months = 15)
}
\arguments{
\item{date}{Date in month to end the display}

\item{months}{past months to include in plot}
}
\value{
a 12 panel plot showing various statistics over months requested: unique members, average workday hours, average daily attendance, new members, supports, members with supports, pre-employment supports, transitional employment supports, supported employment supports, independent employment supports, donors, amount raised
}
\description{
Plot Membership by the Numbers
}
\examples{
plot_membership_by_the_numbers()
}
