% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_supports_by_member_and_type.R
\name{plot_supports_by_member_and_type}
\alias{plot_supports_by_member_and_type}
\title{Plot Supports by Member and Type}
\usage{
plot_supports_by_member_and_type(date = Sys.Date())
}
\arguments{
\item{date}{The date on which the supports should be plotted}
}
\value{
a plot showing the members and for each member, the supports they received as a stacked bar chart
}
\description{
Plot Supports by Member and Type
}
\examples{
plot_supports_by_member_and_type("2023-03-09")
}
