% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_supports_by_staff_and_type.R
\name{plot_supports_by_staff_and_type}
\alias{plot_supports_by_staff_and_type}
\title{Plot Supports by Staff and Type}
\usage{
plot_supports_by_staff_and_type(date = Sys.Date())
}
\arguments{
\item{date}{The date on which the supports should be plotted}
}
\value{
a plot showing the Staff members and for each Staff member, the supports they authored as a stacked bar chart
}
\description{
Plot Supports by Staff and Type
}
\examples{
plot_supports_by_staff_and_type("2023-03-09")
}
