% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top_10_campaigns_by_gifts.R
\name{plot_top_10_campaigns_by_gifts}
\alias{plot_top_10_campaigns_by_gifts}
\title{Plot Top 10 Campaigns by Gifts}
\usage{
plot_top_10_campaigns_by_gifts()
}
\value{
a donut plot with share as fraction of gifts in each of top 10 campaigns
}
\description{
Plot Top 10 Campaigns by Gifts
}
\examples{
plot_top_10_campaigns_by_gifts()
}
