% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_total_raised_by_payment_method.R
\name{plot_total_raised_by_payment_method}
\alias{plot_total_raised_by_payment_method}
\title{Plot total raised all time by payment method}
\usage{
plot_total_raised_by_payment_method()
}
\value{
a plot showing vertical bars for each payment method
}
\description{
Plot total raised all time by payment method
}
\examples{
plot_total_raised_by_payment_method()
}
