% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flourish_connection.R
\name{flourish_connection}
\alias{flourish_connection}
\title{Flourish Connection}
\usage{
flourish_connection()
}
\value{
a database connection to Flourish
}
\description{
Flourish Connection
}
\examples{
con <- flourish_connection()
RMySQL::dbDisconnect(con)
}
