% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_families.R
\name{list_families}
\alias{list_families}
\title{List Family Contact Information}
\usage{
list_families()
}
\value{
a flextable with all families and their contact information
}
\description{
List Family Contact Information
}
\examples{
list_families()
}
