% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.total.raised.by.year.R
\name{plot_total_raised_by_year}
\alias{plot_total_raised_by_year}
\title{Plot total raised by calendar year}
\usage{
plot_total_raised_by_year(date = Sys.Date())
}
\arguments{
\item{date}{Show gifts up to to the end of the year preceding the date}
}
\value{
a plot showing total raised each year in vertical bars
}
\description{
Plot total raised by calendar year
}
\examples{
plot_total_raised_by_year()
}
